requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
django-phonenumber-field-8.0.0
aiohttp-retry==2.8.3
aiosignal==1.3.2
asgiref==3.8.1
attrs==24.3.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.3.1
django-browser-reload==1.17.0
python-decouple
django-sendsms==0.5
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
incremental==24.7.2
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
propcache==0.2.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
twilio==9.4.3
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.8.2
yarl==1.18.3
zope-interface==7.2
africastalking
channels
django-ckeditor==6.7.2
djangorestframework==3.15.1